Changed Settings View
