+++ src/components/flight/TrackingMap.vue
@@ -1,5 +1,4 @@
 <script setup lang="ts">
-  import type { Flight } from '@/models/Flight'
   import { ref } from 'vue'
 
   import 'leaflet/dist/leaflet.css'
@@ -12,11 +11,13 @@
   import { generateLine } from '@/scripts/flight/tracking/map'
 
   const zoom = ref(2)
+
+  const mapRef = ref()
 </script>
 
 <template>
   <div style="height: 100%; width: 100%">
-    <l-map ref="map" v-model:zoom="zoom" :center="[47.41322, -1.219482]">
+    <l-map ref="mapRef" v-model:zoom="zoom" :center="[47.41322, -1.219482]">
       <l-tile-layer
         url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
         layer-type="base"
