+++ src/components/settings/overlays/AircraftCreationOverlay.vue
