+++ src/scripts/utils/time.ts
@@ -1,0 +1,5 @@
+export const dateToHoursMinutes = (date: Date) => {
+  const hours = date.getHours()
+  const minutes = date.getMinutes()
+  return `${hours}:${minutes}`
+}
